Move hook fixture to conftest for compiled tests
